[Refactor] camera: 카메라뷰 Preview와 BottomContainer로 분리
